# verhub.openapi.yaml
openapi: 3.1.0
info:
  title: Verhub API
  version: 1.2.0
  description: |
    Verhub 后端接口契约（与当前 NestJS Controller + DTO 保持一致）。
    基础路径统一为 /api/v1。
servers:
  - url: /api/v1
    description: 默认 API 前缀
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前管理员信息
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/admin-profile:
    get:
      tags: [Auth]
      summary: 获取管理员资料
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Auth]
      summary: 更新管理员资料
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/password:
    patch:
      tags: [Auth]
      summary: 修改管理员密码
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/account:
    patch:
      tags: [Auth]
      summary: 修改管理员账号（用户名）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminAccountDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: 用户名冲突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/tokens:
    get:
      tags: [Auth]
      summary: 获取长期 Token 列表
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiTokenItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Auth]
      summary: 创建长期 Token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiTokenDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/api-keys:
    get:
      tags: [Auth]
      summary: 获取长期 Token 列表（包装响应）
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Auth]
      summary: 创建长期 Token（包装响应）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiTokenDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/tokens/{id}:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    delete:
      tags: [Auth]
      summary: 撤销长期 Token
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/api-keys/{id}:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    delete:
      tags: [Auth]
      summary: 撤销长期 Token（包装响应）
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/projects/announcements:
    post:
      tags: [Announcements]
      summary: 通过 project_key 创建公告（兼容路径）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectAnnouncementDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/announcements/{announcement_id}:
    parameters:
      - name: announcement_id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Announcements]
      summary: 按公告ID更新公告（兼容路径）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Announcements]
      summary: 按公告ID删除公告（兼容路径）
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/feedbacks/{feedback_id}:
    parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Feedbacks]
      summary: 按反馈ID更新反馈（兼容路径）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Feedbacks]
      summary: 按反馈ID删除反馈（兼容路径）
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /public/{projectKey}/announcements:
    parameters:
      - $ref: "#/components/parameters/ProjectKey"
    get:
      tags: [Announcements]
      summary: 按 project_key 获取公告列表
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/{projectKey}/announcements/latest:
    parameters:
      - $ref: "#/components/parameters/ProjectKey"
    get:
      tags: [Announcements]
      summary: 按 project_key 获取最新公告
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementItem"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/projects/{projectId}/actions:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Actions]
      summary: 获取项目行为分类列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/projects/actions:
    post:
      tags: [Actions]
      summary: 创建项目行为分类
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/actions/{action_id}:
    parameters:
      - name: action_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Actions]
      summary: 获取行为分类下的行为记录
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRecordListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Actions]
      summary: 更新行为分类
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActionDto"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Actions]
      summary: 删除行为分类
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/actions/record/{action_record_id}:
    parameters:
      - name: action_record_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Actions]
      summary: 获取单条行为记录
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRecordItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/actions/statistics:
    get:
      tags: [Actions]
      summary: 获取行为分类统计
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/actions/record/statistics:
    get:
      tags: [Actions]
      summary: 获取行为记录统计
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /public/{projectKey}/actions:
    parameters:
      - $ref: "#/components/parameters/ProjectKey"
    post:
      tags: [Actions]
      summary: 上报行为记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRecordDto"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRecordItem"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: 版本创建接口支持 API Key 认证
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
...
    EntityId:
      name: id
      in: path
      required: true
      schema:
        type: string
    ProjectKey:
      name: projectKey
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: 未认证或令牌非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DeleteSuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
    LoginDto:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 64
        password:
          type: string
          minLength: 6
          maxLength: 128
    LoginResponse:
      type: object
      required: [access_token, expires_in, user]
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          description: JWT 剩余秒数
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      required: [id, username, role, must_change_password]
      properties:
        id:
          type: string
        username:
          type: string
          maxLength: 64
        role:
          type: string
          enum: [ADMIN]
        must_change_password:
          type: boolean
    AdminProfileResponse:
      $ref: "#/components/schemas/AuthUser"
    UpdateAdminProfileDto:
      type: object
      required: [current_password]
      properties:
        current_password:
          type: string
          minLength: 6
          maxLength: 128
        username:
          type: string
          maxLength: 64
        new_password:
          type: string
          minLength: 6
          maxLength: 128
    ChangePasswordDto:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 6
          maxLength: 128
        new_password:
          type: string
          minLength: 6
          maxLength: 128
    UpdateAdminAccountDto:
      type: object
      required: [current_password, username]
      properties:
        current_password:
          type: string
          minLength: 6
          maxLength: 128
        username:
          type: string
          maxLength: 64
    CreateApiTokenDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 64
        scopes:
          type: array
          items:
            type: string
            maxLength: 64
        expires_in_days:
          type: integer
          minimum: 1
          maximum: 365
    ApiTokenListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ApiTokenItem"
    ApiTokenItem:
      type: object
      required: [id, name, scopes, is_active, created_at, expires_at, revoked_at, last_used_at]
      properties:
        id:
          type: string
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
        revoked_at:
          type: [string, "null"]
          format: date-time
        last_used_at:
          type: [string, "null"]
          format: date-time
    CreateApiTokenResponse:
      type: object
      required: [id, name, token, scopes, expires_at, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        token:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ClientPlatform:
      type: string
      enum: [ios, android, windows, mac, web]
...
    CreateAnnouncementDto:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 128
        content:
          type: string
          maxLength: 4096
        is_pinned:
          type: boolean
    CreateProjectAnnouncementDto:
      allOf:
        - $ref: "#/components/schemas/CreateAnnouncementDto"
        - type: object
          required: [project_key]
          properties:
            project_key:
              type: string
              maxLength: 64
    UpdateAnnouncementDto:
      type: object
      properties:
        title:
          type: string
          maxLength: 128
        content:
          type: string
          maxLength: 4096
        is_pinned:
          type: boolean
    AnnouncementItem:
      type: object
      required: [id, title, content, is_pinned, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        is_pinned:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnnouncementListResponse:
      type: object
      required: [total, data]
      properties:
        total:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/AnnouncementItem"
...
    UpdateFeedbackDto:
      type: object
      properties:
        user_id:
          type: string
          maxLength: 64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
          maxLength: 4096
        platform:
          $ref: "#/components/schemas/ClientPlatform"
        custom_data:
          type: object
          additionalProperties: true
    FeedbackItem:
      type: object
      required: [id, user_id, rating, content, platform, custom_data, created_at]
      properties:
        id:
          type: string
        user_id:
          type: [string, "null"]
        rating:
          type: [integer, "null"]
        content:
          type: string
        platform:
          oneOf:
            - $ref: "#/components/schemas/ClientPlatform"
            - type: "null"
        custom_data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        created_at:
          type: string
          format: date-time
...
    CreateActionDto:
      type: object
      required: [project_key, name, description]
      properties:
        project_key:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 512
        custom_data:
          type: object
          additionalProperties: true
    UpdateActionDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 512
        custom_data:
          type: object
          additionalProperties: true
    ActionItem:
      type: object
      required: [action_id, project_key, name, description, custom_data, created_time]
      properties:
        action_id:
          type: string
        project_key:
          type: string
        name:
          type: string
        description:
          type: string
        custom_data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        created_time:
          type: integer
    ActionListResponse:
      type: object
      required: [total, data]
      properties:
        total:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActionItem"
    CreateActionRecordDto:
      type: object
      required: [action_id]
      properties:
        action_id:
          type: string
          maxLength: 64
        custom_data:
          type: object
          additionalProperties: true
    ActionRecordItem:
      type: object
      required: [action_record_id, action_id, created_time, http, custom_data]
      properties:
        action_record_id:
          type: string
        action_id:
          type: string
        created_time:
          type: integer
        http:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        custom_data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
    ActionRecordListResponse:
      type: object
      required: [total, data]
      properties:
        total:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActionRecordItem"